# productCenter/api/sku/v1/sku_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.2
// - protoc             v7.35.0
// source: sku/v1/sku.proto

package v1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationSkuCreateSku = "/sku.v1.Sku/CreateSku"
const OperationSkuDeleteSku = "/sku.v1.Sku/DeleteSku"
const OperationSkuGetSku = "/sku.v1.Sku/GetSku"
const OperationSkuListSkus = "/sku.v1.Sku/ListSkus"
const OperationSkuUpdateSku = "/sku.v1.Sku/UpdateSku"

type SkuHTTPServer interface {
	// CreateSku 创建SKU
	CreateSku(context.Context, *CreateSkuRequest) (*CreateSkuResponse, error)
	// DeleteSku 删除SKU
	DeleteSku(context.Context, *DeleteSkuRequest) (*DeleteSkuResponse, error)
	// GetSku 查看SKU详情
	GetSku(context.Context, *GetSkuRequest) (*GetSkuResponse, error)
	// ListSkus 分页查询SKU列表（支持按商品筛选）
	ListSkus(context.Context, *ListSkusRequest) (*ListSkusResponse, error)
	// UpdateSku 更新SKU
	UpdateSku(context.Context, *UpdateSkuRequest) (*UpdateSkuResponse, error)
}

func RegisterSkuHTTPServer(s *http.Server, srv SkuHTTPServer) {
	r := s.Route("/")
	r.POST("/api/v1/skus", _Sku_CreateSku0_HTTP_Handler(srv))
	r.PUT("/api/v1/skus/{id}", _Sku_UpdateSku0_HTTP_Handler(srv))
	r.GET("/api/v1/skus/{id}", _Sku_GetSku0_HTTP_Handler(srv))
	r.GET("/api/v1/skus", _Sku_ListSkus0_HTTP_Handler(srv))
	r.DELETE("/api/v1/skus/{id}", _Sku_DeleteSku0_HTTP_Handler(srv))
}

func _Sku_CreateSku0_HTTP_Handler(srv SkuHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in CreateSkuRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationSkuCreateSku)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.CreateSku(ctx, req.(*CreateSkuRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*CreateSkuResponse)
		return ctx.Result(200, reply)
	}
}

func _Sku_UpdateSku0_HTTP_Handler(srv SkuHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in UpdateSkuRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationSkuUpdateSku)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.UpdateSku(ctx, req.(*UpdateSkuRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*UpdateSkuResponse)
		return ctx.Result(200, reply)
	}
}

func _Sku_GetSku0_HTTP_Handler(srv SkuHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetSkuRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationSkuGetSku)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetSku(ctx, req.(*GetSkuRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetSkuResponse)
		return ctx.Result(200, reply)
	}
}

func _Sku_ListSkus0_HTTP_Handler(srv SkuHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ListSkusRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationSkuListSkus)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListSkus(ctx, req.(*ListSkusRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ListSkusResponse)
		return ctx.Result(200, reply)
	}
}

func _Sku_DeleteSku0_HTTP_Handler(srv SkuHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in DeleteSkuRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationSkuDeleteSku)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.DeleteSku(ctx, req.(*DeleteSkuRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*DeleteSkuResponse)
		return ctx.Result(200, reply)
	}
}

type SkuHTTPClient interface {
	// CreateSku 创建SKU
	CreateSku(ctx context.Context, req *CreateSkuRequest, opts ...http.CallOption) (rsp *CreateSkuResponse, err error)
	// DeleteSku 删除SKU
	DeleteSku(ctx context.Context, req *DeleteSkuRequest, opts ...http.CallOption) (rsp *DeleteSkuResponse, err error)
	// GetSku 查看SKU详情
	GetSku(ctx context.Context, req *GetSkuRequest, opts ...http.CallOption) (rsp *GetSkuResponse, err error)
	// ListSkus 分页查询SKU列表（支持按商品筛选）
	ListSkus(ctx context.Context, req *ListSkusRequest, opts ...http.CallOption) (rsp *ListSkusResponse, err error)
	// UpdateSku 更新SKU
	UpdateSku(ctx context.Context, req *UpdateSkuRequest, opts ...http.CallOption) (rsp *UpdateSkuResponse, err error)
}

type SkuHTTPClientImpl struct {
	cc *http.Client
}

func NewSkuHTTPClient(client *http.Client) SkuHTTPClient {
	return &SkuHTTPClientImpl{client}
}

// CreateSku 创建SKU
func (c *SkuHTTPClientImpl) CreateSku(ctx context.Context, in *CreateSkuRequest, opts ...http.CallOption) (*CreateSkuResponse, error) {
	var out CreateSkuResponse
	pattern := "/api/v1/skus"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationSkuCreateSku))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// DeleteSku 删除SKU
func (c *SkuHTTPClientImpl) DeleteSku(ctx context.Context, in *DeleteSkuRequest, opts ...http.CallOption) (*DeleteSkuResponse, error) {
	var out DeleteSkuResponse
	pattern := "/api/v1/skus/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationSkuDeleteSku))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "DELETE", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// GetSku 查看SKU详情
func (c *SkuHTTPClientImpl) GetSku(ctx context.Context, in *GetSkuRequest, opts ...http.CallOption) (*GetSkuResponse, error) {
	var out GetSkuResponse
	pattern := "/api/v1/skus/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationSkuGetSku))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// ListSkus 分页查询SKU列表（支持按商品筛选）
func (c *SkuHTTPClientImpl) ListSkus(ctx context.Context, in *ListSkusRequest, opts ...http.CallOption) (*ListSkusResponse, error) {
	var out ListSkusResponse
	pattern := "/api/v1/skus"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationSkuListSkus))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// UpdateSku 更新SKU
func (c *SkuHTTPClientImpl) UpdateSku(ctx context.Context, in *UpdateSkuRequest, opts ...http.CallOption) (*UpdateSkuResponse, error) {
	var out UpdateSkuResponse
	pattern := "/api/v1/skus/{id}"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationSkuUpdateSku))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PUT", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}