# productCenter/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /api/v1/skus:
        get:
            tags:
                - Sku
            description: 分页查询SKU列表（支持按商品筛选）
            operationId: Sku_ListSkus
            parameters:
                - name: page
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: pageSize
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: productId
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/sku.v1.ListSkusResponse'
        post:
            tags:
                - Sku
            description: 创建SKU
            operationId: Sku_CreateSku
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/sku.v1.CreateSkuRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/sku.v1.CreateSkuResponse'
    /api/v1/skus/{id}:
        get:
            tags:
                - Sku
            description: 查看SKU详情
            operationId: Sku_GetSku
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/sku.v1.GetSkuResponse'
        put:
            tags:
                - Sku
            description: 更新SKU
            operationId: Sku_UpdateSku
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/sku.v1.UpdateSkuRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/sku.v1.UpdateSkuResponse'
        delete:
            tags:
                - Sku
            description: 删除SKU
            operationId: Sku_DeleteSku
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/sku.v1.DeleteSkuResponse'
...
components:
    schemas:
...
        sku.v1.CreateSkuRequest:
            type: object
            properties:
                productId:
                    type: string
                sku:
                    type: string
                title:
                    type: string
                price:
                    type: string
                stock:
                    type: string
                imgUrl:
                    type: string
            description: 创建SKU
        sku.v1.CreateSkuResponse:
            type: object
            properties:
                sku:
                    $ref: '#/components/schemas/sku.v1.SkuInfo'
        sku.v1.DeleteSkuResponse:
            type: object
            properties:
                success:
                    type: boolean
        sku.v1.GetSkuResponse:
            type: object
            properties:
                sku:
                    $ref: '#/components/schemas/sku.v1.SkuInfo'
        sku.v1.ListSkusResponse:
            type: object
            properties:
                skus:
                    type: array
                    items:
                        $ref: '#/components/schemas/sku.v1.SkuInfo'
                total:
                    type: integer
                    format: int32
                page:
                    type: integer
                    format: int32
                pageSize:
                    type: integer
                    format: int32
        sku.v1.SkuInfo:
            type: object
            properties:
                id:
                    type: string
                productId:
                    type: string
                sku:
                    type: string
                title:
                    type: string
                price:
                    type: string
                stock:
                    type: string
                imgUrl:
                    type: string
                createdAt:
                    type: string
                updatedAt:
                    type: string
            description: SKU实体
        sku.v1.UpdateSkuRequest:
            type: object
            properties:
                id:
                    type: string
                productId:
                    type: string
                sku:
                    type: string
                title:
                    type: string
                price:
                    type: string
                stock:
                    type: string
                imgUrl:
                    type: string
            description: 更新SKU
        sku.v1.UpdateSkuResponse:
            type: object
            properties:
                sku:
                    $ref: '#/components/schemas/sku.v1.SkuInfo'
...